* Allow usage of `KryptonMessageBoxExtended`
