+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Messagebox/Designer/KryptonMessageBoxConfiguratorActionList.cs
@@ -13,7 +13,7 @@ namespace Krypton.Toolkit.Suite.Extended.Messagebox
     internal class KryptonMessageBoxConfiguratorActionList : DesignerActionList
     {
         #region Instance Fields
-        private readonly KryptonMessageBoxManager _messageBoxConfigurator;
+        private readonly KryptonMessageBoxManagerOld _messageBoxConfigurator;
 
         private readonly IComponentChangeService _service;
         #endregion
@@ -21,7 +21,7 @@ namespace Krypton.Toolkit.Suite.Extended.Messagebox
         #region Identity
         public KryptonMessageBoxConfiguratorActionList(KryptonMessageBoxConfiguratorDesigner owner) : base(owner.Component)
         {
-            _messageBoxConfigurator = owner.Component as KryptonMessageBoxManager;
+            _messageBoxConfigurator = owner.Component as KryptonMessageBoxManagerOld;
 
             _service = (IComponentChangeService)GetService(typeof(IComponentChangeService));
         }
